[YS-495] refactor: 코드 리뷰 반영
